+++ src/app/api-server/src/config.rs
@@ -7,19 +7,26 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+use std::path::PathBuf;
+
 use serde::{Deserialize, Serialize};
 
 /////////////////////////////////////////////////////////////////////////////////////////
 
 #[derive(Debug, Default, PartialEq, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
 #[serde(rename_all = "camelCase")]
 pub struct ApiServerConfig {
     pub auth: AuthConfig,
+    pub repo: RepoConfig,
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////
+// Auth
+/////////////////////////////////////////////////////////////////////////////////////////
 
 #[derive(Debug, Default, PartialEq, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
 #[serde(rename_all = "camelCase")]
 pub struct AuthConfig {
     pub providers: Vec<AuthProviderConfig>,
@@ -28,6 +35,7 @@ pub struct AuthConfig {
 /////////////////////////////////////////////////////////////////////////////////////////
 
 #[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
 #[serde(rename_all = "camelCase")]
 #[serde(tag = "kind")]
 pub enum AuthProviderConfig {
@@ -39,12 +47,36 @@ pub enum AuthProviderConfig {
 
 #[derive(Debug, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
+#[serde(deny_unknown_fields)]
 pub struct AuthProviderConfigGitHub {}
 
 /////////////////////////////////////////////////////////////////////////////////////////
 
 #[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
 #[serde(rename_all = "camelCase")]
 pub struct AuthProviderConfigDummy {
     pub accounts: Vec<kamu::domain::auth::AccountInfo>,
+}
+
+/////////////////////////////////////////////////////////////////////////////////////////
+// Repo
+/////////////////////////////////////////////////////////////////////////////////////////
+
+#[derive(Debug, Default, PartialEq, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
+#[serde(rename_all = "camelCase")]
+pub struct RepoConfig {
+    pub caching: RepoCachingConfig,
+}
+
+#[derive(Debug, Default, PartialEq, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
+#[serde(rename_all = "camelCase")]
+pub struct RepoCachingConfig {
+    // Caches dataset handles to avoid expensive S3 bucket scans
+    pub registry_cache_enabled: bool,
+
+    // Stores metadata blocks in a local directory to avoid many tiny S3 requests
+    pub metadata_local_fs_cache_path: Option<PathBuf>,
 }