+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Support for metadata object caching on local file system (e.g. to avoid too many calls to S3 repo)
+- Support for caching the list of datasets in a remote repo (e.g. to avoid very expensive S3 bucket prefix listing calls)
+
 ## [0.15.1] - 2024-03-14
 ### Fixed
 - OData adapter will ignore fields with unsupported data types instead of crashing